Implement AsyncThunk and EntityAdapter in ScoresList
Similar update to gamesSlice, though the latter has no UI
implementation yet; it remains to change the game
creation pattern so that a game (i.e. cubes) is created
on the client and the data then pushed to the server to
allow the game to persist.

It also remains to implement an addNewScore thunk and create
users (on the server) and implement all game logic.